fixed blindly removing old logfiles
* allows logrotation to append to already existing files
* did further testing, fixes #2124

(cherry picked from commit a94ba9c28642e128daa452fd93f0046c1bf93561)

Conflicts:
	vdemo_base.sh

+++ vdemo_base.sh
@@ -120,6 +120,7 @@ function vdemo_start_component {
 				logfiledir="${VDEMO_logfile%/*}"
 				if [ ! -d "$logfiledir" ]; then mkdir -p "$logfiledir"; fi
 				if [ "$LOG_ROTATION" == "ON" ]; then
+					echo "logrotation is enabled." >&2
 					component_logrotate_configfile=${VDEMO_logfile_prefix}${VDEMO_title}.rotation.conf
 					component_logrotate_statefile=${VDEMO_logfile_prefix}${VDEMO_title}.rotation.state
 
@@ -153,7 +154,9 @@ function vdemo_start_component {
 	done
 
 	export -f component
-	rm -f ${VDEMO_logfile} # remove any previous log file
+	if [ ! "$LOG_ROTATION" == "ON" ]; then
+		rm -f ${VDEMO_logfile} # remove any old log file
+	fi
 
 	test -n "${VDEMO_logging}" && \
 	echo "starting $VDEMO_title with component function:"$'\n'"$(declare -f component)" \
